Refactor code base
- Refactor profile password reset and profile registration based on
Code Climate findings
